Fix the export in CommonJS environments
Summary:
If the current module is used in a mixed environment, loaded as part of
a CommonJS style module, but where the scopechain contains an AMD style
'defined' function, this function would wrongfully use this to define
itself instead of exporting using the module/exports variables.

This diff changes the order of the tests so that it prefers exporting to
the `module` object, then the `exports` object and finally using `define`.

Test Plan: Ran the test suite, and loaded the module inside a local
CommonJS environment running in a page also hosting an AMD style loader.

Reviewers: kitcambridge